+++ example/model/server_error.dart
@@ -1,0 +1,6 @@
+class ServerError {
+  final String message;
+  ServerError(this.message);
+  @override
+  String toString() => "ServerError($message)";
+}