+++ lib/either_option.dart
@@ -2,9 +2,3 @@ library either_option;
 
 export 'src/either.dart';
 export 'src/option.dart';
-
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}