+++ lib/src/option.dart
@@ -3,35 +3,67 @@ import 'package:either_option/src/either.dart';
 /**
  * Simple Option monad implementation 
  */
-
 abstract class Option<A> {
+  /**
+   * Return [None] Option
+   */
   static Option<A> empty<A>() => none();
 
+  /**
+   * Return [None] if null else [Some] of A
+   */
   static Option<A> of<A>(A a) => a != null ? some(a) : none();
 
+  /**
+   * Applies [onNone] if this is a [None] or [onSome] if this is a [Some] of A
+   */
   Z fold<Z>(Z Function() onNone, Z Function(A a) onSome);
 
+  /**
+   * True if Some of A else false
+   */
   bool get isDefined => fold(() => false, (_) => true);
 
+  /**
+   * True if None else false
+   */
   bool get isEmpty => !isDefined;
 
+  /**
+   * Return [a] inside [Some] else  supplied [caseElse] if None
+   */
   A getOrElse(A caseElse) => fold(() => caseElse, (A a) => a);
 
+  /**
+   * Return inchanged Option if [Some] else supplied [caseElse] if None
+   */
   Option<A> orElse(Option<A> caseElse) =>
       fold(() => caseElse, (A a) => this); // or  (A a) => some(a)
 
+  /**
+   * Return Some of Application of [f] on [a] inside [Some] if [isDefined] else None
+   */
   Option<Z> map<Z>(Z Function(A a) f) => fold(none, (A a) => some(f(a)));
 
+  /**
+   * Return Application of [f] on [a] inside [Some] if [isDefined] else `None`
+   */
   Option<Z> flatMap<Z>(Option<Z> Function(A a) f) => fold(none, (A a) => f(a));
 
+  /**
+   * Return [Left] from Option
+   */
   Either<A, A> toLeft(A caseNone) =>
       fold(() => Left(caseNone), (A a) => Left(a));
 
+  /**
+   * Return [Right] from Option
+   */
   Either<A, A> toRight(A caseNone) =>
       fold(() => Right(caseNone), (A a) => Right(a));
 
   /**
-   * Consider that None is a Left(something) and Some(a) is a Right(a)
+   * Consider that None is a [Left](something) and [Some](a) is a Right(a)
    */
   Either<L, A> toEither<L>(L leftValue) =>
       fold(() => Left<L, A>(leftValue), (A a) => Right<L, A>(a));