+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## [0.0.1] - TODO: Add release date.
+# Changelog
 
-* TODO: Describe initial release.
+## [0.0.1] - 2019-07-03
+### Added
+
+- Option Monad implementation
+- Either Monad implementation
+- Units tests
+- Some examples